test(vuls2): fold the CPE rule tables into vuls2_test.go
